# fe-pickup/openapi.yaml
openapi: 3.1.0
info:
  title: PickUp API
  description: PickUp API 문서
  version: v1
servers:
- url: /
...
paths:
...
  /consignments:
    get:
      tags:
      - Consignment
      summary: 내 상품 목록 조회
      description: |
        내가 등록한 상품 목록을 판매 상태(status)로 필터링하여 조회합니다.
        상품 ID 내림차순으로 조회하며, 다음 페이지 조회 시 이전 응답의 cursor를 전달합니다.
      operationId: getMyConsignments
      parameters:
      - name: status
        in: query
        required: true
        schema:
          type: string
          minLength: 1
      - name: cursor
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: size
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: 내 상품 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetMyConsignmentsPageResponse"
              examples:
                내 상품 목록 조회 결과:
                  description: 내 상품 목록 조회 결과
                  value:
                    hasNext: true
                    cursor: 99
                    size: 1
                    items:
                    - consignmentId: 100
                      auctionId: null
                      card:
                        cardId: 10
                        cardName: 리자몽 1st Edition Holo
                        setName: Base Set
                        cardNumber: 4/102
                        language: 일본어
                        rarity: MINT
                        imageUrl: https://example.com/cards/10.png
                      sellerMemberId: 1
                      majorDefect: 모서리에 약간의 마모
                      status: REGISTERABLE
                      auctionStatus: null
                      auctionStartedAt: null
                      auctionEndedAt: null
                      certificate:
                        certificateId: 200
                        serialNumber: PSA-84213907
                        certificationBody: PSA
                        grade: "10"
                        gradeCode: GEM_MINT
                        inspectedAt: 2026-06-30
        "400":
          description: "잘못된 조회 조건 (필수 파라미터 누락, 유효하지 않은 상태 값 등)"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "401":
          description: 인증이 필요함 (access-token 쿠키 없음/만료)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
      security:
      - access-token-cookie: []
    post:
      tags:
      - Consignment
      summary: 상품 등록
      description: |
        카드 검색으로 선택한 카드와 감정서, 이미지(앞/뒤 최소 2장) 정보를 받아
        상품을 등록합니다. 이미지 순서는 요청 배열에 담긴 순서를 그대로 사용합니다.
        등록 직후 상태는 REGISTERABLE(등록 가능)로 설정됩니다.
      operationId: registerConsignment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterConsignmentRequest"
            examples:
              상품 등록 요청:
                description: 상품 등록 요청
                value:
                  cardId: 10
                  majorDefect: 모서리에 약간의 마모
                  certificate:
                    serialNumber: PSA-84213907
                    certificationBody: PSA
                    grade: "10"
                    inspectedAt: 2026-06-30
                  images:
                  - temporaryObjectKey: uploads/1/consignments/00000000-0000-0000-0000-000000000001.jpg
                  - temporaryObjectKey: uploads/1/consignments/00000000-0000-0000-0000-000000000002.jpg
        required: true
      responses:
        "201":
          description: 상품 등록 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RegisterConsignmentResponse"
              examples:
                상품 등록 결과:
                  description: 상품 등록 결과
                  value:
                    consignmentId: 100
                    card:
                      cardId: 10
                      cardName: 리자몽 1st Edition Holo
                      setName: Base Set
                      cardNumber: 4/102
                      language: 일본어
                      rarity: MINT
                      imageUrl: https://example.com/cards/10.png
                    sellerMemberId: 1
                    majorDefect: 모서리에 약간의 마모
                    status: REGISTERABLE
                    certificate:
                      certificateId: 200
                      serialNumber: PSA-84213907
                      certificationBody: PSA
                      grade: "10"
                      gradeCode: GEM_MINT
                      inspectedAt: 2026-06-30
        "400":
          description: "요청 값 검증 실패 (필수 값 누락, 이미지 2장 미만, 유효하지 않은 등급/감정기관 등)"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "401":
          description: 인증이 필요함 (access-token 쿠키 없음/만료)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "404":
          description: 카드 또는 회원을 찾을 수 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "409":
          description: 이미 등록된 인증서 일련번호
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
      security:
      - access-token-cookie: []
...
  /consignments/{consignmentId}:
    get:
      tags:
      - Consignment
      summary: 상품 상세 조회
      description: 상품 ID로 상품 상세 정보를 조회합니다.
      operationId: getConsignment
      parameters:
      - name: consignmentId
        in: path
        description: 상품 ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: 상품 상세 조회 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GetConsignmentDetailResponse"
              examples:
                상품 상세 조회 결과:
                  description: 상품 상세 조회 결과
                  value:
                    consignmentId: 100
                    card:
                      cardId: 10
                      cardName: 리자몽 1st Edition Holo
                      setName: Base Set
                      cardNumber: 4/102
                      language: 일본어
                      rarity: MINT
                      imageUrl: https://example.com/cards/10.png
                    sellerMemberNickname: 피카츄
                    majorDefect: 모서리에 약간의 마모
                    status: REGISTERABLE
                    auctionStatus: null
                    auctionStartedAt: null
                    auctionEndedAt: null
                    certificate:
                      certificateId: 200
                      serialNumber: PSA-84213907
                      certificationBody: PSA
                      grade: "10"
                      gradeCode: GEM_MINT
                      inspectedAt: 2026-06-30
                    images:
                    - consignmentImageId: 1
                      imageOrder: 1
                      imageUrl: https://example.com/cards/10-front.png
                    - consignmentImageId: 2
                      imageOrder: 2
                      imageUrl: https://example.com/cards/10-back.png
                    auctionRegistered: false
        "404":
          description: 상품을 찾을 수 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
    delete:
      tags:
      - Consignment
      summary: 상품 삭제
      description: 상품을 삭제합니다. 경매 신청 이후(IN_AUCTION/SOLD) 상태의
        상품은 삭제할 수 없습니다.
      operationId: deleteConsignment
      parameters:
      - name: consignmentId
        in: path
        description: 상품 ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: 상품 삭제 성공
        "401":
          description: 인증이 필요함 (access-token 쿠키 없음/만료)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "403":
          description: 본인이 등록한 상품이 아님
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "404":
          description: 상품을 찾을 수 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "409":
          description: 경매가 시작된 이후 상태라 삭제할 수 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
      security:
      - access-token-cookie: []
    patch:
      tags:
      - Consignment
      summary: 상품 정보 수정
      description: |
        상품의 major defect, 감정서, 이미지 목록을 수정합니다.
        images는 수정 후 남길 최종 이미지 목록입니다. 기존 이미지는 consignmentImageId로 유지하고
        새 이미지는 temporaryObjectKey로 추가하며, 목록에서 빠진 기존 이미지는 삭제합니다.
        배열 순서가 최종 이미지 순서가 됩니다.
        경매 신청 이후(IN_AUCTION)/낙찰 완료(SOLD) 상태에서는 수정할 수 없습니다.
      operationId: modifyConsignment
      parameters:
      - name: consignmentId
        in: path
        description: 상품 ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModifyConsignmentRequest"
            examples:
              상품 정보 수정 요청:
                description: 상품 정보 수정 요청
                value:
                  majorDefect: 모서리에 약간의 마모
                  certificate:
                    serialNumber: PSA-84213907
                    certificationBody: PSA
                    grade: "10"
                    inspectedAt: 2026-06-30
                  images:
                  - consignmentImageId: 1
                  - temporaryObjectKey: uploads/1/consignments/00000000-0000-0000-0000-000000000003.jpg
        required: true
      responses:
        "200":
          description: 상품 정보 수정 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GetConsignmentDetailResponse"
              examples:
                상품 정보 수정 결과:
                  description: 상품 정보 수정 결과
                  value:
                    consignmentId: 100
                    card:
                      cardId: 10
                      cardName: 리자몽 1st Edition Holo
                      setName: Base Set
                      cardNumber: 4/102
                      language: 일본어
                      rarity: MINT
                      imageUrl: https://example.com/cards/10.png
                    sellerMemberNickname: 피카츄
                    majorDefect: 모서리에 약간의 마모
                    status: REGISTERABLE
                    auctionStatus: null
                    auctionStartedAt: null
                    auctionEndedAt: null
                    certificate:
                      certificateId: 201
                      serialNumber: PSA-84213907
                      certificationBody: PSA
                      grade: "10"
                      gradeCode: GEM_MINT
                      inspectedAt: 2026-06-30
                    images:
                    - consignmentImageId: 1
                      imageOrder: 1
                      imageUrl: https://example.com/cards/10-front.png
                    - consignmentImageId: 3
                      imageOrder: 2
                      imageUrl: https://example.com/cards/10-back.png
                    auctionRegistered: false
        "400":
          description: "요청 값 검증 실패 (이미지 2장 미만, 유효하지 않은 등급/감정기관 등)"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "401":
          description: 인증이 필요함 (access-token 쿠키 없음/만료)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "403":
          description: 본인이 등록한 상품이 아님
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "404":
          description: 상품을 찾을 수 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "409":
          description: "경매 신청 이후/진행 중인 상품이라 수정할 수 없거나, 다른 상품이 이미 사용 중인 인증서 일련번호"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
      security:
      - access-token-cookie: []
...
components:
  schemas:
...
    ExceptionResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
...
    CertificateRequest:
      type: object
      properties:
        serialNumber:
          type: string
          minLength: 1
        certificationBody:
          type: string
          minLength: 1
        grade:
          type: string
          minLength: 1
        inspectedAt:
          type: string
          format: date
      required:
      - certificationBody
      - grade
      - inspectedAt
      - serialNumber
    ConsignmentImageRequest:
      type: object
      properties:
        consignmentImageId:
          type: integer
          format: int64
        temporaryObjectKey:
          type: string
    RegisterConsignmentRequest:
      type: object
      properties:
        cardId:
          type: integer
          format: int64
        majorDefect:
          type: string
        certificate:
          $ref: "#/components/schemas/CertificateRequest"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ConsignmentImageRequest"
          maxItems: 5
          minItems: 2
      required:
      - cardId
      - certificate
      - images
...
    RegisterConsignmentResponse:
      type: object
      properties:
        consignmentId:
          type: integer
          format: int64
        card:
          $ref: "#/components/schemas/SearchCardsResponse"
        sellerMemberId:
          type: integer
          format: int64
        majorDefect:
          type: string
        status:
          type: string
          enum:
          - REGISTERABLE
          - IN_AUCTION
          - SOLD
        certificate:
          $ref: "#/components/schemas/CertificateResponse"
    SearchCardsResponse:
      type: object
      properties:
        cardId:
          type: integer
          format: int64
        cardName:
          type: string
        setName:
          type: string
        cardNumber:
          type: string
        language:
          type: string
        rarity:
          type: string
        imageUrl:
          type: string
...
    ModifyConsignmentRequest:
      type: object
      properties:
        majorDefect:
          type: string
        certificate:
          $ref: "#/components/schemas/CertificateRequest"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ConsignmentImageRequest"
          maxItems: 5
          minItems: 2
      required:
      - certificate
      - images
...
    GetConsignmentDetailResponse:
      type: object
      properties:
        consignmentId:
          type: integer
          format: int64
        card:
          $ref: "#/components/schemas/GetCardDetailResponse"
        sellerMemberNickname:
          type: string
        majorDefect:
          type: string
        status:
          type: string
          enum:
          - REGISTERABLE
          - IN_AUCTION
          - SOLD
        auctionStatus:
          type: string
          enum:
          - SCHEDULED
          - ONGOING
          - WON
          - PASSED
        auctionStartedAt:
          type: string
          format: date-time
        auctionEndedAt:
          type: string
          format: date-time
        certificate:
          $ref: "#/components/schemas/CertificateResponse"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ConsignmentImageResponse"
        auctionRegistered:
          type: boolean
    GetMyConsignmentsResponse:
      type: object
      properties:
        consignmentId:
          type: integer
          format: int64
        auctionId:
          type: integer
          format: int64
        card:
          $ref: "#/components/schemas/GetCardDetailResponse"
        sellerMemberId:
          type: integer
          format: int64
        majorDefect:
          type: string
        status:
          type: string
          enum:
          - REGISTERABLE
          - IN_AUCTION
          - SOLD
        auctionStatus:
          type: string
          enum:
          - SCHEDULED
          - ONGOING
          - WON
          - PASSED
        auctionStartedAt:
          type: string
          format: date-time
        auctionEndedAt:
          type: string
          format: date-time
        certificate:
          $ref: "#/components/schemas/CertificateResponse"
    GetMyConsignmentsPageResponse:
      type: object
      properties:
        hasNext:
          type: boolean
        cursor:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: "#/components/schemas/GetMyConsignmentsResponse"
...
  securitySchemes:
    access-token-cookie:
      type: apiKey
      description: 로그인 시 발급되는 access-token 쿠키
      name: access-token
      in: cookie